+++ tests/Pest.php
@@ -1,0 +1,7 @@
+<?php
+
+declare(strict_types=1);
+
+use NotificationChannels\Infobip\Test\TestCase;
+
+uses(TestCase::class)->in(__DIR__);